requirements.txt
grpcio 
grpcio-tools
google
google-api-python-client
opencv-python 
tensorflow 
onnx 
mxnet
tqdm 
torch 
scikit-image
Pillow
unique-id
faiss-cpu
dotmap
imutils
unidecode
torchvision
onnxruntime
asyncio-mqtt
schedule
getmac
matplotlib
websockets
python-dotenv